+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 > Features:
+- [84f7f3c] [Feature] In-chat model management
 - [c10571d] [Feature] Add Mistral Medium 3
 - [f5136d5] [Feature] OpenAI Responses API
 - [ac8ec65] [Feature] Add Alibaba qwen-vl models
@@ -8,9 +9,6 @@
 
 > Bug Fixes:
 - [6d1599b] [Bugfix] Handle unhandled handle_with_exceptions
-
-> Others:
-- [accc3b2] [Other] Patch markdownify due to vulnerability as per Dependabot #3
 
 
 ---